+++ src/main/java/com/api/ttoklip/domain/town/community/post/repository/CommunityRepositoryImpl.java
@@ -1,11 +1,11 @@
 package com.api.ttoklip.domain.town.community.post.repository;
 
 import static com.api.ttoklip.domain.member.domain.QMember.member;
-import static com.api.ttoklip.domain.privacy.domain.QProfile.profile;
 import static com.api.ttoklip.domain.town.community.comment.QCommunityComment.communityComment;
 import static com.api.ttoklip.domain.town.community.image.entity.QCommunityImage.communityImage;
 import static com.api.ttoklip.global.util.SecurityUtil.getCurrentMember;
 
+import com.api.ttoklip.domain.profile.domain.QProfile;
 import com.api.ttoklip.domain.town.TownCriteria;
 import com.api.ttoklip.domain.town.community.comment.CommunityComment;
 import com.api.ttoklip.domain.town.community.post.entity.Community;
@@ -28,6 +28,7 @@ public class CommunityRepositoryImpl implements CommunityRepositoryCustom {
     private static final String SPLIT_CRITERIA = " ";
     private final JPAQueryFactory jpaQueryFactory;
     private final QCommunity community = QCommunity.community;
+    private final QProfile profile = QProfile.profile;
 
     @Override
     public Community findByIdActivated(final Long communityId) {
