+++ src/main/java/com/api/ttoklip/domain/town/community/post/repository/CommunitySearchRepository.java
@@ -1,10 +1,10 @@
 package com.api.ttoklip.domain.town.community.post.repository;
 
 import static com.api.ttoklip.domain.member.domain.QMember.member;
-import static com.api.ttoklip.domain.privacy.domain.QProfile.profile;
 import static com.api.ttoklip.domain.town.community.like.entity.QCommunityLike.communityLike;
 import static com.api.ttoklip.domain.town.community.scrap.entity.QCommunityScrap.communityScrap;
 
+import com.api.ttoklip.domain.profile.domain.QProfile;
 import com.api.ttoklip.domain.town.community.comment.QCommunityComment;
 import com.api.ttoklip.domain.town.community.post.entity.Community;
 import com.api.ttoklip.domain.town.community.post.entity.QCommunity;
@@ -28,6 +28,8 @@ import org.springframework.util.StringUtils;
 public class CommunitySearchRepository {
 
     private final JPAQueryFactory jpaQueryFactory;
+
+    private final QProfile profile = QProfile.profile;
 
     private final QCommunity community = QCommunity.community;
