+++ src/main/java/com/api/ttoklip/global/upload/Uploader.java
@@ -1,0 +1,11 @@
+package com.api.ttoklip.global.upload;
+
+import java.util.List;
+import org.springframework.web.multipart.MultipartFile;
+
+public interface Uploader {
+
+    String uploadMultipartFile(MultipartFile multipartFile);
+
+    List<String> uploadMultipartFiles(List<MultipartFile> multipartFiles);
+}
